editing category added

+++ src/components/Button/IconButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './IconButton.css';
 
 interface IconButtonProps {
-  type: 'delete' | 'add' | 'remove'; // Button type
+  type: 'delete' | 'add' | 'remove' | 'confirm' | 'edit' | 'save' ; // Button type
   onClick: () => void; // Click handler
   label?: string; // Optional label for accessibility
 }
@@ -13,6 +13,9 @@ const IconButton: React.FC<IconButtonProps> = ({ type, onClick, label }) => {
       {type === 'delete' && 'x'}
       {type === 'add' && '+'}
       {type === 'remove' && '-'}
+      {type === 'confirm' && '✅'}
+      {type === 'edit' && '✏️'}
+      {type === 'save' && '💾'}
     </button>
   );
 };
